requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1

# NLP & ML
spacy==3.7.2
scispacy==0.5.4
sentence-transformers==2.2.2
transformers==4.36.0
torch>=2.0.0

# Medical NER models (will be downloaded via spacy)
# en_core_sci_md @ https://s3-us-west-2.amazonaws.com/ai2-s3-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3.tar.gz

# NLI and embeddings
rank-bm25==0.2.2

# LLM APIs
google-generativeai==0.3.2
requests==2.31.0

# Data processing
datasets==2.15.0
pandas==2.1.4
numpy==1.26.2

# Storage
duckdb==0.9.2

# Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Utilities
tqdm==4.66.1
tenacity==8.2.3
colorama==0.4.6

# Development
pytest==7.4.3
black==23.12.0

